move enums to .const
