+++ homeassistant/components/tolo/select.py
@@ -14,8 +14,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from . import ToloSaunaCoordinatorEntity, ToloSaunaUpdateCoordinator
-from .const import DOMAIN
-from .enums import AromaTherapySlot, LampMode
+from .const import DOMAIN, AromaTherapySlot, LampMode
 
 
 @dataclass(frozen=True, kw_only=True)
